+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5]
+
+### Added
+
+- Added more comprehensive docs in README.md
+
+### Removed
+
+- Removed `cluster_autoscaling` as it only applies to a clusters default node pool which are not supported by this module.
+
 ## [0.0.4]
 
 ### Added
@@ -70,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Implementation
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.4...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mineiros-io/terraform-google-gke-cluster/compare/v0.0.1...v0.0.2